Made adjustment on the filter as requested.
